builder/expr: out with the old, in with the new
This push is apart of the past few weeks of work that involve an expression system capable of both
quick serialization and linq expression generation. The idea was to separate the LINQ expressions
from the script builder such that these scripts could be easily serialized and re-built directly
from our own custom expression tree structures.